requirements.txt
# -------------------------------------
# Web Framework and API
# -------------------------------------
fastapi==0.109.2
uvicorn==0.22.0
pydantic==2.6.4
pydantic-settings==2.2.1

# -------------------------------------
# Database
# -------------------------------------
sqlalchemy==2.0.29
aiosqlite==0.19.0
alembic==1.12.0
greenlet==1.1.3 

# -------------------------------------
# Kubernetes Integration
# -------------------------------------
kubernetes==26.1.0

# -------------------------------------
# Machine Learning and Data Processing
# -------------------------------------
tensorflow==2.18.0
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
scipy==1.13.0
joblib==1.4.2

# -------------------------------------
# Visualization (Optional, used in training/analysis)
# -------------------------------------
matplotlib==3.10.0
seaborn==0.13.2

# -------------------------------------
# AI Integration
# -------------------------------------
google-genai

# -------------------------------------
# Monitoring and Metrics
# -------------------------------------
prometheus-client==0.16.0
httpx==0.24.1

# -------------------------------------
# Logging and Configuration
# -------------------------------------
loguru==0.7.2

# -------------------------------------
# Utils
# -------------------------------------
python-multipart==0.0.6
aiohttp==3.9.5
tenacity==8.2.3
requests==2.31.0
